quotes/views.py
# This is a placeholder views.py file for the quotes app.
# Actual views should be implemented based on AGENTS.md and README.md.
from rest_framework import generics
# from .models import Quote
# from .serializers import QuoteSerializer

# Example views (to be defined based on actual requirements):

# For /api/designs/{design_id}/quotes/
# class DesignQuoteListCreateView(generics.ListCreateAPIView):
#     serializer_class = QuoteSerializer
#     # permission_classes = [IsAuthenticated, IsDesignOwnerOrManufacturer]

#     def get_queryset(self):
#         # design_id = self.kwargs['design_id']
#         # return Quote.objects.filter(design_id=design_id)
#         return [] # Placeholder

#     def perform_create(self, serializer):
#         # design = Design.objects.get(pk=self.kwargs['design_id'])
#         # serializer.save(manufacturer=self.request.user, design=design)
#         pass # Placeholder

# For /api/quotes/{quote_id}/
# class QuoteDetailView(generics.RetrieveUpdateDestroyAPIView):
#     queryset = Quote.objects.all()
#     serializer_class = QuoteSerializer
#     # permission_classes = [IsAuthenticated, IsQuoteOwnerOrManufacturer]
#     lookup_field = 'quote_id' # or 'pk' if using default
