requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.0
cffi==1.17.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
boto3>=1.34.0 # Added boto3 for AWS S3 interaction
celery>=5.3.0 # For asynchronous tasks
redis>=5.0.0   # For Celery broker (and potentially result backend)
numpy>=1.20.0 # Dependency for numpy-stl
numpy-stl>=2.17.0 # For STL file analysis
steputils==0.1  # For basic STEP file interactions (IGES support pending suitable library)
# igesutils>=0.1.3  # For basic IGES file interactions - Installation failed
# python-occ-core>=7.7.0 # For CAD analysis (OpenCASCADE wrapper) - Installation failed
